Moved cred.js into lib.
