+++ src/lib/cred.js
